Use config as initial value, more product work
